Refactor RTI and invert handshake
