+++ simaas/rti/default.py
@@ -2,88 +2,91 @@ import asyncio
 import json
 import os
 import socket
-import threading
 import time
 
 import traceback
-from threading import Lock
 from typing import Optional, List, Tuple, Dict
 
-from fastapi import Request
-
-from sqlalchemy import create_engine, Column, String
-from sqlalchemy.orm import sessionmaker, declarative_base
-from sqlalchemy_json import NestedMutableJson
+import docker
+from docker.models.containers import Container
 
 from simaas.cli.cmd_proc_builder import clone_repository, build_processor_image
 from simaas.cli.cmd_rti import shorten_id
-from simaas.core.helpers import generate_random_string, get_timestamp_now
-from simaas.core.identity import Identity
+from simaas.core.helpers import get_timestamp_now
 from simaas.core.logging import Logging
 from simaas.core.schemas import GithubCredentials
 from simaas.dor.protocol import P2PLookupDataObject, P2PFetchDataObject
-from simaas.helpers import docker_find_image, docker_load_image, docker_delete_image, docker_run_job_container, \
-    docker_kill_job_container, docker_container_running
+from simaas.helpers import docker_find_image, docker_load_image, docker_delete_image
 from simaas.p2p.base import P2PAddress
-from simaas.p2p.protocol import P2PLatency
-from simaas.rti.api import RTIService
-from simaas.rti.exceptions import RTIException, ProcessorNotDeployedError
-from simaas.rti.protocol import P2PInterruptJob, P2PRunnerPerformHandshake, P2PPushJob
-from simaas.rti.schemas import Processor, Job, Task, JobStatus
+from simaas.rti.base import RTIServiceBase, DBDeployedProcessor, DBJobInfo
+from simaas.rti.exceptions import RTIException
+from simaas.rti.protocol import P2PInterruptJob
+from simaas.rti.schemas import Processor, Job
 from simaas.dor.schemas import GitProcessorPointer, DataObject, ProcessorDescriptor
 
 logger = Logging.get('rti.service')
 
-Base = declarative_base()
+
+def docker_run_job_container(image_name: str, p2p_address: Tuple[str, int],
+                             custodian_address: str, custodian_pubkey: str, job_id: str) -> str:
+    client = docker.from_env()
+    container = client.containers.run(
+        image=image_name,
+        volumes={
+            # job_path: {'bind': '/job', 'mode': 'rw'}
+        },
+        ports={
+            '6000/tcp': p2p_address,
+        },
+        detach=True,
+        stderr=True, stdout=True,
+        auto_remove=False,
+        environment={
+            'SIMAAS_CUSTODIAN_ADDRESS': custodian_address,
+            'SIMAAS_CUSTODIAN_PUBKEY': custodian_pubkey,
+            'JOB_ID': job_id
+        }
+    )
+
+    return container.id
 
 
-class DBDeployedProcessor(Base):
-    __tablename__ = 'deployed_processor'
-    id = Column(String(64), primary_key=True)
-    state = Column(String, nullable=False)
-    image_name = Column(String, nullable=True)
-    gpp = Column(NestedMutableJson, nullable=True)
-    error = Column(String, nullable=True)
+def docker_kill_job_container(container_id: str) -> None:
+    client = docker.from_env()
+    container = client.containers.get(container_id)
+    container.kill()
 
 
-class DBJobInfo(Base):
-    __tablename__ = 'job_info'
-    id = Column(String(64), primary_key=True)
-    proc_id = Column(String(64), nullable=False)
-    user_iid = Column(String(64), nullable=False)
-    p2p_address_pub = Column(String(64), nullable=False)
-    p2p_address_sec = Column(String(64), nullable=False)
-    status = Column(NestedMutableJson, nullable=False)
-    job = Column(NestedMutableJson, nullable=False)
-    container_id = Column(String(16), nullable=False)
-    peer = Column(NestedMutableJson, nullable=True)
+def docker_delete_container(container_id: str) -> None:
+    client = docker.from_env()
+    container = client.containers.get(container_id)
+    container.remove()
 
 
-class DefaultRTIService(RTIService):
+def docker_container_running(container_id: str) -> bool:
+    client = docker.from_env()
+    container = client.containers.get(container_id)
+    return container.status == "running"
+
+
+def docker_container_list() -> Dict[str, Container]:
+    client = docker.from_env()
+    return {
+        container.id: container for container in client.containers.list()
+    }
+
+
+class DefaultRTIService(RTIServiceBase):
     def __init__(self, node, db_path: str, retain_job_history: bool = False, strict_deployment: bool = True):
-        super().__init__(retain_job_history=retain_job_history, strict_deployment=strict_deployment)
+        super().__init__(
+            node=node, db_path=db_path, retain_job_history=retain_job_history, strict_deployment=strict_deployment
+        )
 
         # initialise properties
-        self._mutex = Lock()
-        self._node = node
         self._port_range = (6000, 9000)
         self._most_recent_port = None
 
-        # initialise directories
-        self._jobs_path = os.path.join(self._node.datastore, 'jobs')
-        self._procs_path = os.path.join(self._node.datastore, 'procs')
-        logger.info(f"[init] using jobs path at {self._jobs_path}")
-        logger.info(f"[init] using procs path at {self._procs_path}")
-        os.makedirs(self._jobs_path, exist_ok=True)
-        os.makedirs(self._procs_path, exist_ok=True)
-
-        # initialise database things
-        logger.info(f"[init] using database at {db_path}")
-        self._engine = create_engine(db_path)
-        Base.metadata.create_all(self._engine)
-        self._session_maker = sessionmaker(bind=self._engine)
-
-    def _perform_deployment(self, proc: Processor) -> None:
+    def perform_deploy(self, proc: Processor) -> None:
         loop = asyncio.new_event_loop()
         try:
             # search the network for the processor docker image data object and fetch it
@@ -186,7 +189,7 @@ class DefaultRTIService(RTIService):
         finally:
             loop.close()
 
-    def _perform_undeployment(self, proc_id: str, image_name: str, keep_image: bool = True) -> None:
+    def perform_undeploy(self, proc_id: str, image_name: str, keep_image: bool = True) -> None:
         # remove the record from the db
         with self._mutex:
             with self._session_maker() as session:
@@ -206,52 +209,75 @@ class DefaultRTIService(RTIService):
                 trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
                 logger.error(f"[undeploy:{shorten_id(proc_id)}] failed to delete docker image {image_name}: {trace}")
 
-    def _perform_job_initialise(self, p2p_address_pub: str, p2p_address_sec: str, job: Job, container_id: str) -> None:
+    def perform_submit(self, proc: Processor, job: Job) -> dict:
+        # determine P2P addresses
+        p2p_addresses = self._find_available_addresses(n=1)
+        runner_p2p_address: Tuple[str, int] = p2p_addresses[0]
+
+        # start the job container and keep the container id
+        logger.info(f"[submit:{shorten_id(proc.id)}] [job:{job.id}] start job container")
+        container_id = docker_run_job_container(
+            proc.image_name, runner_p2p_address, self._node.p2p.address(), self._node.identity.c_public_key, job.id
+        )
+
+        # update the runner information
+        with self._session_maker() as session:
+            # get the record and status
+            record = session.query(DBJobInfo).get(job.id)
+            record.runner['container_id'] = container_id
+            session.commit()
+
+        return {
+            'container_id': container_id
+        }
+
+    def perform_cancel(self, job_id: str, peer_address: P2PAddress, grace_period: int = 30) -> None:
+        # attempt to cancel the job
         try:
-            custodian = self._node
-
-            logger.info(f"[job:{job.id}] BEGIN initialising job...")
-
-            # wait until the socket can be reached
-            latency, attempt = asyncio.run(P2PLatency.perform_unsecured(p2p_address_pub))
-            logger.info(f"[job:{job.id}] needed {attempt} attempts to test latency: {latency} msec")
-
-            # perform the handshake
-            response: Tuple[GitProcessorPointer, Identity] = asyncio.run(P2PRunnerPerformHandshake.perform(
-                p2p_address_pub, custodian.keystore, custodian.p2p.address()
-            ))
-            peer: Identity = response[1]
-
-            # update the peer information in the database
-            with self._session_maker() as session:
-                record = session.query(DBJobInfo).get(job.id)
-                record.peer = peer.model_dump()
-                session.commit()
-
-            # create P2P address
-            peer_address = P2PAddress(
-                address=p2p_address_sec,
-                curve_secret_key=custodian.keystore.curve_secret_key(),
-                curve_public_key=custodian.keystore.curve_public_key(),
-                curve_server_key=peer.c_public_key
-            )
-
-            # wait until the socket can be reached
-            latency, attempt = asyncio.run(P2PLatency.perform(peer_address))
-            logger.info(f"[job:{job.id}] needed {attempt} attempts to test latency: {latency} msec")
-
-            # push the job to the runner -> this will trigger execution
-            asyncio.run(P2PPushJob.perform(peer_address, job))
-
-            logger.info(f"[job:{job.id}] END initialising job -> OK")
+            asyncio.run(P2PInterruptJob.perform(peer_address))
 
         except Exception as e:
             trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
-            logger.info(f"[job:{job.id}] END initialising job -> FAILED: {trace}")
+            logger.warning(f"[job:{job_id}] attempt to cancel job {job_id} failed: {trace}")
+            grace_period = 0
 
-            # something went wrong, kill the container to avoid zombies
-            docker_kill_job_container(container_id)
+        with self._session_maker() as session:
+            # get the record and status
+            record = session.query(DBJobInfo).get(job_id)
+            container_id = record.runner['container_id']
 
+            deadline = get_timestamp_now() + grace_period * 1000
+            while get_timestamp_now() < deadline:
+                try:
+                    # sleep for a bit and check the record
+                    time.sleep(1)
+
+                    # is the container still running -> if not, all good we are done
+                    if not docker_container_running(container_id):
+                        return
+
+                except Exception as e:
+                    trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
+                    logger.warning(f"[job:{job_id}] checking status failed: {trace}")
+                    break
+
+            # if we get here then either the deadline has been reached or there was an exception -> kill container
+            try:
+                logger.warning(f"[job:{job_id}] grace period exceeded -> "
+                               f"killing Docker container {container_id}")
+                docker_kill_job_container(container_id)
+
+            except Exception as e:
+                trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
+                logger.warning(f"[job:{job_id}] killing Docker container {container_id} failed: {trace}")
+
+    def perform_purge(self, record: DBJobInfo) -> None:
+        # try to kill the container (if anything is left)
+        try:
+            docker_kill_job_container(record.container_id)
+        except Exception as e:
+            trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
+            logger.warning(f"[job:{record.id}] killing Docker container {record.container_id} failed: {trace}")
 
     def _find_available_addresses(self, n: int = 2, max_attempts: int = 100) -> List[Tuple[str, int]]:
         def is_port_free(_address: Tuple[str, int]) -> bool:
@@ -287,392 +313,3 @@ class DefaultRTIService(RTIService):
                 return addresses
 
         raise RuntimeError("No free ports found in the specified range.")
-
-    def is_deployed(self, proc_id: str) -> bool:
-        with self._session_maker() as session:
-            record = session.query(DBDeployedProcessor).get(proc_id)
-            return record is not None
-
-    def get_all_procs(self) -> List[Processor]:
-        """
-        Retrieves a dict of all deployed processors by their id.
-        """
-        with self._session_maker() as session:
-            records = session.query(DBDeployedProcessor).all()
-            result = []
-            for record in records:
-                result.append(Processor(id=record.id, state=Processor.State(record.state),
-                                        image_name=record.image_name,
-                                        gpp=GitProcessorPointer.model_validate(record.gpp) if record.gpp else None,
-                                        error=record.error))
-
-            return result
-
-    def get_proc(self, proc_id: str) -> Optional[Processor]:
-        """
-        Retrieves a specific processors given its id.
-        """
-        with self._session_maker() as session:
-            record = session.query(DBDeployedProcessor).get(proc_id)
-            if record:
-                return Processor(id=record.id, state=Processor.State(record.state),
-                                 image_name=record.image_name,
-                                 gpp=GitProcessorPointer.model_validate(record.gpp) if record.gpp else None,
-                                 error=record.error)
-            else:
-                return None
-
-    def deploy(self, proc_id: str) -> Processor:
-        """
-        Deploys a processor.
-        """
-
-        # is the processor already deployed?
-        proc = self.get_proc(proc_id)
-        if proc is not None:
-            return proc
-
-        # begin deployment
-        with self._mutex:
-            # create a placeholder processor object
-            proc = Processor(id=proc_id, state=Processor.State.BUSY_DEPLOY, image_name=None, gpp=None, error=None)
-
-            # update or create db record
-            with self._session_maker() as session:
-                record = session.query(DBDeployedProcessor).get(proc.id)
-                if record:
-                    record.state = proc.state.value
-                    record.image_name = proc.image_name
-                    record.gpp = proc.gpp.model_dump()
-                    record.error = proc.error
-
-                else:
-                    session.add(DBDeployedProcessor(id=proc.id, state=proc.state.value, image_name=proc.image_name,
-                                                    gpp=proc.gpp.model_dump() if proc.gpp else None, error=None))
-                session.commit()
-
-            # start the deployment worker
-            threading.Thread(target=self._perform_deployment, args=(proc,)).start()
-
-            return proc
-
-    def undeploy(self, proc_id: str) -> Optional[Processor]:
-        """
-        Removes a processor from the RTI (if it exists).
-        """
-        with self._mutex:
-            with self._session_maker() as session:
-                # do we have a db record for this processor?
-                record = session.query(DBDeployedProcessor).get(proc_id)
-                if not record:
-                    return None
-
-                # create the processor object
-                proc = Processor(id=record.id, state=Processor.State(record.state),
-                                 image_name=record.image_name,
-                                 gpp=GitProcessorPointer.model_validate(record.gpp) if record.gpp else None,
-                                 error=record.error)
-
-                # is the state failed? -> delete the db record
-                if proc.state == Processor.State.FAILED:
-                    logger.warning(f"[undeploy:{shorten_id(proc_id)}] processor failed -> removing it. "
-                                   f"error: {record.error}")
-                    session.delete(record)
-                    session.commit()
-
-                # is the state ready? -> update state to 'busy' and begin undeployment
-                elif proc.state == Processor.State.READY:
-                    # update the state to busy
-                    proc.state = Processor.State.BUSY_UNDEPLOY
-                    record.state = proc.state.value
-                    session.commit()
-
-                    # start the worker
-                    threading.Thread(target=self._perform_undeployment, args=(proc.id, proc.image_name,)).start()
-
-                # is the state busy going up? -> throw error
-                elif proc.state == Processor.State.BUSY_DEPLOY:
-                    raise RTIException("Cannot undeploy a processor that is currently deploying. Try again later.")
-
-                # is the state busy going down? -> do nothing
-                elif proc.state == Processor.State.BUSY_DEPLOY:
-                    logger.warning(f"[undeploy:{shorten_id(proc_id)}] already undeploying.")
-
-                return proc
-
-    def submit(self, proc_id: str, task: Task, request: Request) -> Job:
-        """
-        Submits a task to a deployed processor, thereby creating a new job. Authorisation is required by the owner
-        of the task/job.
-        """
-
-        # get the user's identity and check if it's identical with that's indicated in the task
-        iid = request.headers['saasauth-iid']
-        if iid != task.user_iid:
-            raise RTIException("Mismatch between user indicated in task and user making request", details={
-                'iid': iid,
-                'task': task
-            })
-        user: Identity = self._node.db.get_identity(iid)
-
-        # get the processor
-        proc = self.get_proc(task.proc_id)
-        if proc is None:
-            raise ProcessorNotDeployedError({
-                'proc_id': proc_id
-            })
-
-        # create the job folder with a generated job id
-        job_id = generate_random_string(8)
-        job_path = os.path.join(self._jobs_path, job_id)
-        os.makedirs(job_path, exist_ok=True)
-
-        # create the initial job descriptor and write to file
-        job = Job(id=job_id, task=task, retain=self._retain_job_history, custodian=self._node.info,
-                  proc_name=proc.gpp.proc_descriptor.name, t_submitted=get_timestamp_now())
-        descriptor_path = os.path.join(job_path, 'job.descriptor')
-        with open(descriptor_path, 'w') as f:
-            # noinspection PyTypeChecker
-            json.dump(job.model_dump(), f, indent=2)
-
-        # determine P2P addresses
-        p2p_addresses = self._find_available_addresses(n=2)
-        p2p_address_pub: Tuple[str, int] = p2p_addresses[0]
-        p2p_address_sec: Tuple[str, int] = p2p_addresses[1]
-
-        # create initial job status and write to file
-        status = JobStatus(state=JobStatus.State.UNINITIALISED, progress=0, output={}, notes={},
-                           errors=[], message=None)
-        status_path = os.path.join(job_path, 'job.status')
-        with open(status_path, 'w') as f:
-            # noinspection PyTypeChecker
-            json.dump(status.model_dump(), f, indent=2)
-
-        # start the job container and keep the container id
-        logger.info(f"[submit:{shorten_id(proc_id)}] [job:{job.id}] start job container")
-        container_id = docker_run_job_container(proc.image_name, p2p_address_pub, p2p_address_sec)
-
-        p2p_address_pub: str = f"tcp://{p2p_address_pub[0]}:{p2p_address_pub[1]}"
-        p2p_address_sec: str = f"tcp://{p2p_address_sec[0]}:{p2p_address_sec[1]}"
-
-        with self._session_maker() as session:
-            # create initial job info record and write to database
-            record = DBJobInfo(id=job.id, proc_id=proc_id, user_iid=user.id,
-                               p2p_address_pub=p2p_address_pub,
-                               p2p_address_sec=p2p_address_sec,
-                               status=status.model_dump(), job=job.model_dump(),
-                               container_id=container_id, peer=None)
-            session.add(record)
-            session.commit()
-
-        # start the worker
-        threading.Thread(
-            target=self._perform_job_initialise,
-            args=(p2p_address_pub, p2p_address_sec, job, container_id)
-        ).start()
-
-        return job
-
-    def jobs_by_proc(self, proc_id: str) -> List[Job]:
-        """
-        Retrieves a list of active jobs processed by a processor. Any job that is pending execution or actively
-        executed will be included in the list.
-        """
-        # get the records
-        with self._mutex:
-            with self._session_maker() as session:
-                records = session.query(DBJobInfo).filter_by(proc_id=proc_id).all()
-
-        # parse the records
-        result: List[Job] = []
-        for record in records:
-            status = JobStatus.model_validate(record.status)
-            if status.state in [JobStatus.State.UNINITIALISED, JobStatus.State.INITIALISED, JobStatus.State.RUNNING]:
-                job = Job.model_validate(record.job)
-                result.append(job)
-
-        return result
-
-    def jobs_by_user(self, request: Request) -> List[Job]:
-        """
-        Retrieves a list of active jobs by a user. If the user is the node owner, all active jobs will be returned.
-        """
-        # get the records
-        user: Identity = self._node.db.get_identity(request.headers['saasauth-iid'])
-        with self._mutex:
-            with self._session_maker() as session:
-                if self._node.identity.id == user.id:
-                    records = session.query(DBJobInfo).all()
-                else:
-                    records = session.query(DBJobInfo).filter_by(user_iid=user.id).all()
-
-        # any time period provided?
-        result: List[Job] = []
-        if 'period' in request.query_params:
-            # collect all jobs within the time period
-            cutoff = get_timestamp_now() - int(request.query_params['period']) * 3600 * 1000
-            for record in records:
-                # within time period?
-                job = Job.model_validate(record.job)
-                if job.t_submitted > cutoff:
-                    result.append(job)
-
-        else:
-            # collect ony active jobs
-            for record in records:
-                status = JobStatus.model_validate(record.status)
-                if status.state in [JobStatus.State.UNINITIALISED, JobStatus.State.INITIALISED,
-                                    JobStatus.State.RUNNING]:
-                    job = Job.model_validate(record.job)
-                    result.append(job)
-
-        return result
-
-    def update_job_status(self, job_id: str, job_status: JobStatus) -> None:
-        """
-        Updates the status of a particular job. Authorisation is required by the owner of the job
-        (i.e., the user that has created the job by submitting the task in the first place).
-        """
-        with self._mutex:
-            with self._session_maker() as session:
-                # get the record
-                record = session.query(DBJobInfo).get(job_id)
-                if record is None:
-                    raise RTIException(f"Job {job_id} does not exist.")
-
-                # TODO: might need to think about restricting status updates from jobs that are marked as finished.
-                # # check the status
-                # status = JobStatus.model_validate(record.status)
-                # if status.state not in [JobStatus.State.UNINITIALISED, JobStatus.State.INITIALISED,
-                #                         JobStatus.State.RUNNING]:
-                #     logger.warning(f"Job {job_id} is not active -> status cannot be updated.")
-
-                # update the status
-                record.status = job_status.model_dump()
-                session.commit()
-
-                # update the local status file
-                try:
-                    status_path = os.path.join(self._jobs_path, job_id, 'job.status')
-                    with open(status_path, 'w') as f:
-                        # noinspection PyTypeChecker
-                        json.dump(job_status.model_dump(), f, indent=2)
-                except Exception:
-                    logger.warning(f"Could not write job status to {status_path}")
-
-    def get_job_owner_iid(self, job_id: str) -> str:
-        with self._mutex:
-            with self._session_maker() as session:
-                record = session.query(DBJobInfo).get(job_id)
-                if record is None:
-                    raise RTIException(f"Job {job_id} does not exist.")
-                return record.user_iid
-
-    def get_job_status(self, job_id: str) -> JobStatus:
-        """
-        Retrieves detailed information about the status of a job. Authorisation is required by the owner of the job
-        (i.e., the user that has created the job by submitting the task in the first place).
-        """
-        # get the record
-        with self._mutex:
-            with self._session_maker() as session:
-                record = session.query(DBJobInfo).get(job_id)
-                if record is None:
-                    raise RTIException(f"Job {job_id} does not exist.")
-
-        return JobStatus.model_validate(record.status)
-
-    def _perform_cancel(self, job_id: str, peer_address: P2PAddress, grace_period: int = 30) -> None:
-        # attempt to cancel the job
-        try:
-            asyncio.run(P2PInterruptJob.perform(peer_address))
-
-        except Exception as e:
-            trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
-            logger.warning(f"[job:{job_id}] attempt to cancel job {job_id} failed: {trace}")
-            grace_period = 0
-
-        with self._session_maker() as session:
-            # get the record and status
-            record = session.query(DBJobInfo).get(job_id)
-            container_id = record.container_id
-
-            deadline = get_timestamp_now() + grace_period * 1000
-            while get_timestamp_now() < deadline:
-                try:
-                    # sleep for a bit and check the record
-                    time.sleep(1)
-
-                    # is the container still running -> if not, all good we are done
-                    if not docker_container_running(container_id):
-                        return
-
-                except Exception as e:
-                    trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
-                    logger.warning(f"[job:{job_id}] checking status failed: {trace}")
-                    break
-
-            # if we get here then either the deadline has been reached or there was an exception -> kill container
-            try:
-                logger.warning(f"[job:{job_id}] grace period exceeded -> "
-                               f"killing Docker container {container_id}")
-                docker_kill_job_container(container_id)
-
-            except Exception as e:
-                trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
-                logger.warning(f"[job:{job_id}] killing Docker container {container_id} failed: {trace}")
-
-    def job_cancel(self, job_id: str) -> JobStatus:
-        """
-        Attempts to cancel a running job. Depending on the implementation of the processor, this may or may not be
-        possible.
-        """
-        # get the record
-        with self._mutex:
-            with self._session_maker() as session:
-                record = session.query(DBJobInfo).get(job_id)
-                if record is None:
-                    raise RTIException(f"Job {job_id} does not exist.")
-
-        # check the status
-        status = JobStatus.model_validate(record.status)
-        if status.state not in [JobStatus.State.UNINITIALISED, JobStatus.State.INITIALISED, JobStatus.State.RUNNING]:
-            raise RTIException(f"Job {job_id} is not active -> job cannot be cancelled.")
-
-        # start the cancellation worker
-        peer = Identity.model_validate(record.peer)
-        peer_address = P2PAddress(
-            address=record.p2p_address_sec,
-            curve_secret_key=self._node.keystore.curve_secret_key(),
-            curve_public_key=self._node.keystore.curve_public_key(),
-            curve_server_key=peer.c_public_key
-        )
-        threading.Thread(target=self._perform_cancel, args=(job_id, peer_address)).start()
-
-        return status
-
-    def job_purge(self, job_id: str) -> JobStatus:
-        """
-        Purges a running job. It will be removed regardless of its state.
-        """
-        # remove the job from database
-        with self._mutex:
-            with self._session_maker() as session:
-                # get the record
-                record: Optional[DBJobInfo] = session.query(DBJobInfo).get(job_id)
-                if record is None:
-                    raise RTIException(f"Job {job_id} does not exist.")
-
-                # try to kill the container (if anything is left)
-                try:
-                    docker_kill_job_container(record.container_id)
-                except Exception as e:
-                    trace = ''.join(traceback.format_exception(None, e, e.__traceback__))
-                    logger.warning(f"[job:{job_id}] killing Docker container {record.container_id} failed: {trace}")
-
-                # delete the record
-                session.delete(record)
-                session.commit()
-
-                status = JobStatus.model_validate(record.status)
-                return status
